try handling big diffs

+++ table_delta.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import itertools
 from dataclasses import dataclass
+import time
 
 @dataclass
 class TableSchema:
@@ -85,6 +86,7 @@ def write_diff_to_file(out, table, diff_dict):
 
 
 def process_table_delta(file_today, file_yesterday):
+    file_renamed = False
     dir_today = os.environ['DIR_TODAY']
 
     hosts0 = open(file_today, "r")
@@ -96,12 +98,14 @@ def process_table_delta(file_today, file_yesterday):
         table = TableSchema('', [])
         sub1 = 'COPY'
         sub2 = 'FROM STDIN'
+        skip_file = False
         with open(dir_today + '/' + '_delta.sql', "w") as out:
             out.write('BEGIN;\n')
             out.write('SET client_encoding TO \'UTF8\';\n')
             out.write('SET synchronous_commit TO off;\n')
             b1 = []
             b2 = []
+            start = time.time()
             print('diffing file chunks...')
             for n, lines in enumerate(itertools.zip_longest(hosts0, hosts1)):
                 if not schema_loaded and sub1 in lines[0] and sub2 in lines[0]:
@@ -123,17 +127,32 @@ def process_table_delta(file_today, file_yesterday):
                     b1 = []
                     b2 = []
                     print('moving to next chunk...')
-            b2 = ['' if v is None else v for v in b2]
-            print('diffing remainder...')
-            preprocess_chunk(b1, b2, diff_dict)
-            print('preparing to write to file...')
-            write_diff_to_file(out, table, diff_dict)
-            out.write('COMMIT;\n')
-            os.rename(dir_today + '/' + '_delta.sql', dir_today + '/' + table.name + '_delta.sql')
-            print('delta ready...')
+                    end = time.time()
+                    delta_time = end - start
+                    print(delta_time)
+                    if delta_time > file_max_time:
+                        skip_file = True
+                        break
+            if not skip_file:
+                b2 = ['' if v is None else v for v in b2]
+                print('diffing remainder...')
+                preprocess_chunk(b1, b2, diff_dict)
+                print('preparing to write to file...')
+                write_diff_to_file(out, table, diff_dict)
+                out.write('COMMIT;\n')
+                os.rename(dir_today + '/' + '_delta.sql', dir_today + '/' + table.name + '_delta.sql')
+                print('delta ready...')
+            else:
+                print('delta calculation is taking too long, skipping file, using base file instead')
+                shutil.copyfile(file_today, file_yesterday)
+                os.rename(file_today, dir_today + '/' + table.name + '_delta.sql')
+                os.remove(os.path.join(dir_today,'_delta.sql'))
+                file_renamed = True
+
     finally:
         hosts0.close()
         hosts1.close()
+        return file_renamed
 
 
 if __name__ == '__main__':
@@ -142,13 +161,17 @@ if __name__ == '__main__':
     dir_yesterday = os.environ['DIR_YESTERDAY']
     clean_dirs = int(os.getenv('CLEAN_DIRS', 0))
     chunk_size = int(os.getenv('CHUNK_SIZE', 100000))
+    file_max_time = int(os.getenv('FILE_MAX_TIME', 600))
 
     for i in os.listdir(dir_today):
         if os.path.isfile(os.path.join(dir_today,i)) and '_output.sql' in i:
             print('starting...')
             print(i)
-            process_table_delta(os.path.join(dir_today,i), os.path.join(dir_yesterday,i))
-            if clean_dirs == 1:
-                os.remove(os.path.join(dir_yesterday,i))
-                shutil.move(os.path.join(dir_today,i), os.path.join(dir_yesterday,i))
+            ret = process_table_delta(os.path.join(dir_today,i), os.path.join(dir_yesterday,i))
+            if not ret:
+                if clean_dirs == 1:
+                    print("cleaning up files ...")
+                    os.remove(os.path.join(dir_yesterday,i))
+                    shutil.move(os.path.join(dir_today,i), os.path.join(dir_yesterday,i))
             print('processed...')
+            print('------------------------------------------')
